update to both mkinitramfs and script_header_joetoo with all luks management in the latter now

+++ dev-util/mkinitramfs/scratch/reset-qemu-test.sh
@@ -3,9 +3,14 @@
 source /usr/sbin/script_header_joetoo
 
 base="/home/joe/myUtilities/dev-util/mkinitramfs/scratch"
-message_n "moving to ${base}"
+
+notice_msg_n "moving to ${base}"
 cd "${base}"
 handle_result $? "$PWD" "$PWD"
 
-[ -d initramfs ] && rm -r "initramfs"
+[ -z "$base" ] && { error_msg "error: base is unset"; exit 1; }
+
+find ./ -type d -iname 'initramfs*' -exec rm -rf {} \;
+
 find ./boot -type f -delete
+find ./boot -type l -delete
